Rust FFI: slipstream behind a cargo feature with a stub ABI flavor
